service/ocr-service/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pika==1.3.2
minio>=7.1.0
certifi>=2024.4.0
python-dotenv>=1.0.0
pydantic>=2.1.1
requests>=2.31.0
python-json-logger>=2.0.7

# Database / ORM
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9.0

# Image handling
Pillow>=9.0.0

# Optional
prometheus-client>=0.17.0
